+++ base/src/api/index.rs
@@ -3,7 +3,7 @@ use iron::status;
 
 
 pub fn index(_: &mut Request) -> IronResult<Response> {
-    Ok(Response::with((status::Ok, "API endpoints mounted under /api/v1/")))
+    Ok(Response::with((status::Ok, "Replicante Agent API endpoints")))
 }
 
 
@@ -21,6 +21,6 @@ mod tests {
         ).unwrap();
         let result_body = response::extract_body_to_bytes(response);
         let result_body = String::from_utf8(result_body).unwrap();
-        assert_eq!(result_body, "API endpoints mounted under /api/v1/");
+        assert_eq!(result_body, "Replicante Agent API endpoints");
     }
 }