+++ base/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Introduce an `/api/unstable` API "version".
+- Threads introspection API.
+
+### Changed
+- **BREAKING**: Rename incorrectly named v1 API as unstable.
 
 ## [0.3.0] - 2019-03-29
 ### Changed